+++ tests/tests.cpp
@@ -1,1 +1,2 @@
 #include "search_tests.hpp"
+#include "sort_tests.hpp"