+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+## [0.0.2] 2017-10-19
+
+### Added
+
+- Find smallest algorithm
+- Find largest algorithm
+- Selection sort algorithm
+
 ## [0.0.1] 2017-10-18
 
 ### Added